refactor(dashboard-admin/detail-view): แยก helpers/hooks ออกจาก DetailView โดยไม่เปลี่ยน UI
- Extract โค้ดใน `components/DetailView.tsx` ออกเป็นโมดูลย่อย เพื่อให้ไฟล์คอมโพเนนต์เบา/อ่านง่ายขึ้น และคงพฤติกรรมเดิมทุกจุด
  - helpers/format.ts: fmtDate, safe field readers, coalesceName, gradeConfidence
  - helpers/status.ts: mapStatus, computeNewStatusForAction, actionToApiStatus, STATUS_BADGE_CLASS
  - helpers/mappers.ts: fromApiToDetailVM (+ computeBankNameMatch)
  - hooks/useDetailView.ts: จัดการ optimistic status, confirm flow (approve/reject/override), PATCH ผ่าน proxy, emit "kyc:status-updated"
  - hooks/useDockFooter.ts: แยก logic dock footer จาก effect เดิม

- เพิ่ม types เฉพาะหน้าจอ
  - types/detail.ts: UiStatus, ConfirmKind, DataLogVM, DetailVM

- อัปเดต `components/DetailView.tsx`
  - ใช้ helpers/hooks/types ใหม่ทั้งหมด แต่คง JSX, className, layout, ข้อความ, icon, และคอมเมนต์เดิม (ไม่แตะ UI)
  - คง debug logs เดิมและโครงแท็บ/แบนเนอร์/ฟุตเตอร์เหมือนเดิม 1:1

- เคลียร์โค้ดที่ซ้ำ/ไม่ได้ใช้จากไฟล์เดิม (import/ฟังก์ชันที่ย้ายแล้ว) โดยไม่ลบคอมเมนต์เดิม

fix(types): รองรับ RefObject ที่เป็น nullable ใน useDockFooter เพื่อแก้ TS2345

- เปลี่ยนพารามิเตอร์ของ `useDockFooter` เป็น `React.RefObject<HTMLDivElement | null>`
